+++ src/main/java/fr/oltruong/teamag/entity/Member.java
@@ -31,7 +31,7 @@ public class Member implements Serializable {
     private MemberType memberType;
 
     @Column(nullable = false)
-    private Float estimatedWorkDays;
+    private Double estimatedWorkDays;
 
 
     private String comment;
@@ -109,11 +109,11 @@ public class Member implements Serializable {
         this.memberType = memberType;
     }
 
-    public Float getEstimatedWorkDays() {
+    public Double getEstimatedWorkDays() {
         return estimatedWorkDays;
     }
 
-    public Float getEstimatedWorkMonths() {
+    public Double getEstimatedWorkMonths() {
         return estimatedWorkDays / TeamagConstants.MONTH_DAYS_RATIO;
     }
 
@@ -122,10 +122,10 @@ public class Member implements Serializable {
     }
 
     public String getEstimatedWorkMonthsString() {
-        return String.valueOf((Float) estimatedWorkDays / TeamagConstants.MONTH_DAYS_RATIO);
+        return String.valueOf((Double) estimatedWorkDays / TeamagConstants.MONTH_DAYS_RATIO);
     }
 
-    public void setEstimatedWorkDays(Float estimatedWorkDays) {
+    public void setEstimatedWorkDays(Double estimatedWorkDays) {
         this.estimatedWorkDays = estimatedWorkDays;
     }
